test: complete testing for book in/box, book out/box, onsite verification, check in/out

+++ app/src/main/java/com/etag/stsyn/ui/screen/settings/SettingsScreen.kt
@@ -118,7 +118,7 @@ fun SettingsScreen(
             title = "Power",
             value = appConfiguration.power.toString(),
             description = appConfiguration.power.toString(),
-            onUpdateClick = { viewModel.updateAppConfig(appConfiguration.copy(power = it.toLong())) }
+            onUpdateClick = { viewModel.updateAppConfig(appConfiguration.copy(power = it.toInt())) }
         )
 
         SettingItem(
